refactor: avoid delete already deleted messages (avoids unwanted api calls)

+++ src/utils/deleteMessage.ts
@@ -1,7 +1,8 @@
 import { CustomContext } from "types/customContext.type";
 
-async function deleteMessage(ctx: CustomContext, messageId: number) {
+async function deleteMessage(ctx: CustomContext, messageId: number | null) {
   try {
+    if (!messageId) return;
     await ctx.deleteMessage(messageId);
   } catch (error) {}
 }
